states ^_^

+++ api/v1/views/states.py
@@ -28,7 +28,10 @@ def get_state(state_id):
         obj = storage.get(State, state_id)
     except KeyError:
         abort(404)
-    return jsonify(obj.to_dict())
+    if obj is not None:
+        return jsonify(obj.to_dict())
+    else:
+        abort(404)
 
 
 @app_views.route(
@@ -42,15 +45,21 @@ def delete_state(state_id):
         obj = storage.get(State, state_id)
     except KeyError:
         abort(404)
-    storage.delete(obj)
-    storage.save()
-    return jsonify({}), 200
+    if obj:
+        storage.delete(obj)
+        storage.save()
+        return jsonify({}), 200
+    else:
+        abort(404)
 
 
 @app_views.route("/states", methods=['POST'], strict_slashes=False)
 def new_state():
     """add new state endpoint"""
-    data = request.get_json()
+    if request.content_type == "application/json":
+        data = request.get_json()
+    else:
+        abort(400, 'Not a JSON')
     if data is None:
         abort(400, 'Not a JSON')
     if data.get('name') is None:
@@ -64,15 +73,20 @@ def new_state():
 @app_views.route('/states/<state_id>', methods=['PUT'], strict_slashes=False)
 def update_state(state_id):
     """update state object"""
+    if request.content_type == "application/json":
+        data = request.get_json()
+    else:
+        abort(400, 'Not a JSON')
     try:
         state = storage.get(State, state_id)
     except KeyError:
         abort(404)
-    data = request.get_json()
+    if state is None:
+        abort(404)
     if data is None:
         abort(400, 'Not a JSON')
     for key, value in data.items():
         if key not in ['id', 'created_at', 'updated_at']:
-            state.key = value
+            setattr(state, key, value)
     storage.save()
     return jsonify(state.to_dict()), 200
